+++ persistence/0060_2.7.3_tag_name_length.sql
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS
     attribute INTEGER NOT NULL check(attribute BETWEEN 0 AND 15),
     precision INTEGER NOT NULL check(precision BETWEEN 0 AND 17),
     decimal REAL NOT NULL,
-    offsets REAL NOT NULL,
+    offsets INTEGER NOT NULL,
     type INTEGER NOT NULL check(type BETWEEN 0 AND 19),
     description TEXT NULL check(length(description) <= 512),
     value TEXT,
